+++ src/TRegx/CleanRegex/Replace/ReplacePattern.php
@@ -26,10 +26,10 @@ class ReplacePattern
 
     public function with(string $replacement): string
     {
-        return $this->withRaw(ReplaceReferences::quote($replacement));
+        return $this->withReferences(ReplaceReferences::quote($replacement));
     }
 
-    public function withRaw(string $replacement): string
+    public function withReferences(string $replacement): string
     {
         return preg::replace($this->pattern->pattern, $replacement, $this->subject, $this->limit);
     }